+++ Example/Example/InputBarStyleSelectionController.swift
@@ -7,10 +7,13 @@
 //
 
 import UIKit
+import SwiftUI
 
 class InputBarStyleSelectionController: UITableViewController {
     
     let styles = InputBarStyle.allCases
+
+    let tabBarExampleIndexPath = IndexPath(row: 6, section: 2)
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -30,35 +33,44 @@ class InputBarStyleSelectionController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        if section == 0 {
-            return "InputBarViewController"
-        }
-        return section == 1 ? "InputAccessoryView" : "Subview"
+        return [
+            0: "InputBarViewController",
+            1: "InputAccessoryView",
+            2: "Subview",
+            3: "FAQ/Community Examples"
+        ][section]!
     }
     
     override func numberOfSections(in tableView: UITableView) -> Int {
-        return 3
+        return 4
     }
     
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if section == 0 {
-            return 1
+        switch section {
+        case 0:     return 1
+        case 1...2: return styles.count
+        case 3:     return 3
+        default:    fatalError("unknown section \(section)")
         }
-        return styles.count
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = UITableViewCell()
-        if indexPath.section == 0 {
-            cell.textLabel?.text = "README Preview"
-        } else {
-            cell.textLabel?.text = styles[indexPath.row].rawValue
+        switch (indexPath.section, indexPath.row) {
+        case (0, _):        cell.textLabel?.text = "README Preview"
+        case (1...2, _):    cell.textLabel?.text = styles[indexPath.row].rawValue
+        case (3, 0):        cell.textLabel?.text = "Tab bar example (Slack style)"
+        case (3, 1):        cell.textLabel?.text = "Send button animations"
+        case (3, 2):        cell.textLabel?.text = "SwiftUI example"
+        default:            assertionFailure("unrecognized \(indexPath). Are you trying to add an additional example?")
         }
+
         cell.accessoryType = .disclosureIndicator
         return cell
     }
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+
         if indexPath.section == 0 {
             navigationController?.pushViewController(READMEPreviewViewController(), animated: true)
         } else {
@@ -73,6 +85,22 @@ class InputBarStyleSelectionController: UITableViewController {
                     SubviewExampleViewController(style: styles[indexPath.row],
                                                  conversation: convo),
                     animated: true)
+            } else if indexPath.section == 3 {
+                switch indexPath.row {
+                case 0:
+                    let tabBarController = UITabBarController()
+                    let contained = SubviewExampleViewController(style: InputBarStyle.slack, conversation: convo)
+                    tabBarController.viewControllers = [contained]
+                    navigationController?.pushViewController(tabBarController, animated: true)
+                case 1:
+                    let example = ButtonAnimationExample(style: .imessage, conversation: convo)
+                    navigationController?.pushViewController(example, animated: true)
+                case 2:
+                    let example = UIHostingController(rootView: SwiftUIExample.make(style: .imessage, conversation: convo))
+                    navigationController?.pushViewController(example, animated: true)
+                default:
+                    fatalError("Unknown row \(indexPath.row) in Community Examples section. Are you trying to add a new example?")
+                }
             }
         }
     }