+++ CHANGELOG.md
@@ -1,4 +1,25 @@
 #  Changelog
+- Master:
+- 5.4.0
+   - Make sure framework is ready for XCode 13
+   - Fix availability in AppExtensions
+   - Fix Package.swift to support iOS 12+ only
+   - Reset cachedNotification when keyboard is programmatically dismissed
+- 5.3.1
+   - Applied `additionalBottomSpace` to interactive dismissal
+- 5.3.0
+   - Allow setting `canBecomeFirstResponder`
+   - Fix interactive keyboard dismissal lag
+   - Add example implemntation of `sendButton` animation
+- 5.2.3
+   - Fixed appendSpaceOnCompletion inserting space at wrong location
+- 5.2.2
+   - Added an optional offset in KeyboardManager.bind(tableView:)
+   - Change reuseIdentifier from `public` to `open` to allow inheritance
+   - Fix send button loading indicator for dark mode
+   - Fix iOS 14 UIPasteboard system notification with images
+- 5.2.1
+   - Fix SPM warnings about Info.plist file
 - 5.2.0
     - Drop support for iOS 11 and bump minimum version to iOS 12+
     - Support Swift 5.3 and higher for XCode 12